pyproject.toml
[tool.poetry]
name = "thesis--news-preparation"
version = "0.1.0"
description = "Analyzes and prepares collected news data from ARD Teletext and from Spiegel Online."
authors = ["Michael Wellner"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
scikit-learn = "^0.24.1"
pandas = "^1.2.3"
spacy = "^3.0.3"
text-preprocessing = "^0.1.0"
pyfunctional = "^1.4.3"
tensorflow = "2.4.1"
joblib = "^1.0.1"
langdetect = "^1.0.8"
gensim = "^4.0.0"
python-Levenshtein = "^0.12.2"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
